feat: 게임 내부 flow 소켓 연결 준비
